rename property related with the test case configuration (RoundHasFinalized)
