+++ code/go/0chain.net/conductor/cases/round_has_finalized.go
@@ -3,6 +3,7 @@ package cases
 import (
 	"context"
 	"errors"
+	"fmt"
 	"strconv"
 )
 
@@ -69,6 +70,8 @@ func (n *RoundHasFinalized) AddResult(blob []byte) error {
 	if err := roundInfo.Decode(blob); err != nil {
 		return err
 	}
+
+	fmt.Printf("adding result -> %+v %+v\n", roundInfo, n.roundID)
 
 	if roundInfo.Num == int64(n.roundID) && (n.res == nil || n.res.Num == 0) {
 		defer func() {
