+++ code/go/0chain.net/conductor/conductor/executor.go
@@ -865,7 +865,7 @@ func (r *Runner) ConfigureTestCase(configurator cases.TestCaseConfigurator) erro
 			state.CheckChallengeIsValid = cfg
 
 		case *cases.RoundHasFinalized:
-			state.RoundHasFinalized = cfg
+			state.RoundHasFinalizedConfig = cfg
 
 		default:
 			log.Panicf("unknown test case name: %s", configurator.Name())
