+++ code/go/0chain.net/conductor/conductrpc/state.go
@@ -66,7 +66,7 @@ type State struct {
 	MinerNotarisedBlockRequestor          *cases.MinerNotarisedBlockRequestor
 	FBRequestor                           *cases.FBRequestor
 	MissingLFBTicket                      *cases.MissingLFBTickets
-	RoundHasFinalized                     *cases.RoundHasFinalized
+	RoundHasFinalizedConfig               *cases.RoundHasFinalized
 
 	// Blobbers related states
 	StorageTree          *config.Bad // blobber sends bad files/tree responses
