+++ code/go/0chain.net/miner/miner/miner_main.go
@@ -9,9 +9,10 @@ import (
 )
 
 // stubs that does nothing
-func initIntegrationsTests()        {}
-func registerInConductor(id string) {}
-func shutdownIntegrationTests()     {}
+func initIntegrationsTests()           {}
+func registerInConductor(id string)    {}
+func shutdownIntegrationTests()        {}
+func configureIntegrationsTestsFlags() {}
 func readMagicBlock(magicBlockConfig string) (*block.MagicBlock, error) {
 	return chain.ReadMagicBlockFile(magicBlockConfig)
 }
