--- code/go/0chain.net/miner/miner/stub.go
@@ -1,9 +1,0 @@
-//go:build !integration_tests
-// +build !integration_tests
-
-package main
-
-// stubs that does nothing
-func initIntegrationsTests(id string)  {}
-func shutdownIntegrationTests()        {}
-func configureIntegrationsTestsFlags() {}
